requirements.txt
asgiref==3.5.2
Brotli==1.0.9
cffi==1.15.1
cryptography==38.0.1
csscompressor==0.9.5
Django==4.1.2
django-compress-staticfiles==1.0.1b0
django-model-utils==4.2.0
django-x509==1.1
djangorestframework==3.14.0
importlib-metadata==4.13.0
jsonfield==3.1.0
openwisp-utils==1.0.4
Pillow==9.2.0
pycparser==2.21
pyOpenSSL==22.1.0
PyPDF2==2.11.1
python-environ==0.4.54
pytz==2022.5
qrcode==7.3.1
reportlab==3.6.11
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
zipp==3.9.0
